requirements.txt
Flask==3.0.0
pandas==1.3.3
python-docx==1.1.0
flask-cors==3.0.10
